feat: Implement role management and permissions system
Added a new route for creating roles in the application.

Introduced a PermissionSeeder to seed default permissions for various modules.
Enhanced the OrdersLayout component to include required field validations for order forms.
Created a RoleLayout component with a sidebar for managing bank accounts.
Developed forms and tables for user role and permission, including